Update to click v7.1 and fix tests
Click v7.1.1 fixes several bugs, mostly around windows output
handling, and makes many improvements to helptext handling and related
things. Unfortunately, a couple of our tests rely on the exact
contents of the helptext, so the update breaks our testsuite.

Although the CLI will continue to work on click 7.0 at present, update
the requirement to 7.1.1 so that
- we can use any new features we wish
- users get the latest fixes (forced update for their CLI installs)