Remove test case for failure on click<7.1
In pallets/click#1468 , string parameter types (the default) added a
catchall case in which a value which fails to decode is run though
`decode('utf-8', 'replace')`. This succeeds even on things like
latin-1 chars which can't decode into utf-8 . As a result, our test
case which relied on the value being passed through verbatim doesn't
work anymore. However, the case which uses our custom param type (the
mkdir test) still fails as expected.

We may still be able to and want to test our failure modes around
invalid UTF8 bytes like this, but for the purposes of being click 7.1
compatible, just remove the failing test.