Add hidden `--provision` option to get-identities
This lets you call get-identities with provision=True, but it defaults
to False (as it is today).
This is handy for cases where you know a user's real identity name and
want to put them in ACLs or other places which need the name to be
provisioned in Auth.

As these are rare and mostly internal use-cases, make the flag hidden.