Update doc building to publish release asset
The automatic doc building in GH Actions will publish a release asset,
not a workflow artifact. This is necessary in order for the built docs
to be persistent (workflow artifacts expire).

The downside is that if docs.globus.org references a build artifact,
we might need to "backport" any doc fixes into an old release.
However, there should be no harm in doing so, and we aren't promising
externally that "released" docs are immutable.