Limit upgrades to <2.0 when on py2
First, change the `get_versions()` helper to return a 3-tuple:
  upgrade_target, latest, current

This allows for the upgrade target to differ from the actual latest
version. Specifically, when running under python2, filter the versions
we get back such that upgrade_target is the maximal version satisfying
`< 2.0`.

Then modify the `globus version` and `globus update` commands to
handle this new output.

In the case of the version printing, the logic is somewhat refactored
so that we can more easily distinguish the various conditions that
determine what we say. If the latest and upgrade_target versions
match, the behavior is completely unchanged. But if they differ, we
print a generic message about how the latest CLI version isn't
supported by your runtime and you should reinstall.

In the update command, we just add a check for
`current == upgrade_target` after we've checked for `current == latest`.
If it matches, we print a warning message and exit(0).

Once we release this change, we can recommend that users update to the
latest version of the CLI so that `globus update` will not
unintentionally update them to a version incompatible with a py2
runtime. We can work out exactly what we want to say later, after this
releases.