Begin loading API fixture data from files
This defines a few files which contain fixture data. We then load
these data and pass them into the API route registration method. All
of the parameters to the route registration function are exposed
trivially. The files are pretty much just a walk over nested dicts to
plug things in.

Two little bits of cleverness:
- allow use of goep1 and goep2 in the routes via formatting
- use an extra "metadata" key to return data to the caller when data
  is loaded

To explain this latter part, we can return metadata.foo:123 from a
fixture file. And then when loading the fixture, you know that
"foo=123". This works well for checking that value against other parts
of the data in the fixture file.