Remove encoding tests
These tests are not particularly meaningful without being run against
the live API. They've been replicated elsewhere to test against the
API, but it is not the responsibility of the CLI testsuite to do
anything sophisticated here.

Any remaining tests would just ensure that click and requests behave
as advertised, which is not worth the effort.