Convert 'ls' testing to 'responses'
In addition to all of the basics (copying API responses, plugging them
into fixture files, etc), this also involved a little bit of retooling
of the fixture helpers themselves. Because `ls` calls only vary by
query params, but it's way more comfortable to specify query params as
dicts, allow the use of lists of routes, rather than nested dicts.
That allows multiple routes with the same path to be defined in a
single file, and is a particularly good fit for the `ls` testing.