Convert tests of `rm` functionality
Adds new fixture data for these tests, but also remove several tests
which just send different flags to Transfer and validate the results.
Outside of live API testing, these tests don't do much. In order to be
able to test the flag->payload conversions in the future, should there
ever be a bug in these, build out a simple helper that gets the delete
submission body and check it in one case (the `ignore_missing` flag).