Add support for domains to `globus session update`
Using IdentityType to parse the identities, we can accept usernames,
IDs, or domains. We then either handle things for `--all` as before,
or we split things between domains and non-domain identities. Domains
are exclusive with usernames and identity IDs, and are passed via the
`session_required_single_domain` parameter.

A slight refactor makes the implementation simpler, and identity
usernames are checked against the identity set rather than
get_identities results. This lets us catch identities not in the
identity set (in one case) earlier in the authn process.