Update GitHub Actions and default branch name
The branch name change is master -> main . GitHub says they'll
redirect downloads of tarballs after a branch rename, so the change in
`update` (and the impact there) is basically nil.

This got me looking at the workflow file, and I realized that I didn't
clean it up when I did the testsuite refactor. The games with retrying
a flakey testsuite are no longer necessary.