Add `mutex_option_group` decorator
Several commands check a series of options for mutual exclusivity.
This work is always the same, in essence, and removing it to a
decorator simplifies logic in some of the more complex cases (see the
changes in `validate_endpoint_create_and_update_params`).

Because we're now operating in 3.6+ this introduces some type
annotations, which weren't available to us in the past. We can start
making strictly typed internal APIs.

In the normal cases, `mutex_option_group` will just be given the
option string and will deduce the param name and correctly check for
its presence. However, for unusual cases (explicit underlying param
names, etc) there is a `MutexInfo` object which can be populated with
the correct fields/behaviors, and could be extended further in the
future if necessary.