Add mypy to our lints, fix minor linting issues
Introduce mypy with config to be as pedantic as we can be without
causing unnecessary pain. That is, turn on all of the warnings we can
which are either non-issues or easily resolved. Some settings like
check_untyped_defs would be a huge PITA right now.

This identified three minor issues:
1. `site.USER_BASE` is actually an `Optional[str]`, so switch to
   `getuserbase()`
2. mypy wants a type for `version_ns` in setup.py -- `type: ignore`
   that one to keep complexity in setup.py down
3. `reference/_generate.py` changes the type of a variable, so do some
   minor name-munging to fix it