Switch to globus_sdk.tokenstorage and fix tests
This switches us from ConfigObj-backed token storage in
`~/.globus.cfg` to a SQLiteAdapter from globus_sdk.tokenstorage
(available in 3.0a2).

When the tokenstorage is being initialized, it will look to load data
from config and revoke tokens and delete any templated client. It does
this unless the storage file already exists (in which case the
behavior gets disabled in order to be less disruptive to CLI
maintainers working with both versions).

Get the tests passing after this change, including patching which puts
a memory-backed sqlite adapter in place as "the" token storage
adapter. Various tweaks to the tests are needed to get things to pass
on SDK v3.0a2 and are not immediately related to the storage system
change.

One unrelated change to the CLI is included: because `GlobusResponse`
no longer exists, define a `CLIStubResponse` class for wrapping data
to look similar to a `GlobusHTTPResponse`. The pretty-printer then
looks for either this or `GlobusHTTPResponse` and handles them
appropriately.