Update to sdk v3.0a4 and use new paging
This updates the globus-cli to the latest globus-sdk alpha (v3.0a4)
and adapts usage around paginated calls where appropriate.

Primarily, the changes are:
- make calls in the new `client.paginated.foo()` pattern
- wrap paginators so that we can limit with a custom limit and check
  `has_next()`
- convert the recursive ls helper away from inheriting
  `PaginatedResource` (which is now gone)

In cases where `num_results=None` or `num_results=sys.maxsize` were
used, no wrapping is necessary. Simply ask the paginator for `items()`
and allow any number of results to be returned (as determined and
potentially limited by the SDK).

In cases where a `num_results=limit` was set, this is no longer possible.
`limit` is a pagination parameter. A wrapper object applies an
external `limit` on the total number of items returned.
Similarly, the wrapper can easily implement `has_next()` by always
getting one more item than is needed and simply checking if
StopIteration was reached.