Rename 'safeio' to 'termio'
The 'safeio' name originated when earlier versions of click needed to
be wrapped to avoid issues writing to broken pipes, as would often
happen when piping command output to `head` or `tail`.

The click.echo changes in the years since have rendered this name
obsolete. `safeio` doesn't really capture what the package is
responsible for: it's really our way of handling all output to the
terminal, so it's out 'termio' package.