Add explicit control for CLI interactivity
The CLI currently uses a heuristic for determining an interactive
session of "is PS1 set?"

This has proven solid and accurate for many years, in spite of there
not being any standard method for determining if a TTY is interactive.

This introduces an explicit flag which can be used to force
interactive mode on or off (strtobool conversion).
Set `GLOBUS_CLI_INTERACTIVE=0` and you won't get any interactive
behaviors.

As the 3.x line will potentially introduce more auto-prompt behaviors,
this will be useful for testing, and may prove important for some users
as well.