Add login `--gcs` opt and `collection delete`
Update the LoginManager to support dynamic requirements in service of
this new flag.

The LoginManager feature isn't really GCS-specific, but makes the
login manager capable of handling newly-added, dynamic requirements.
Primarily, it adds storage for dynamic requirements and replaces use
of the STATIC_SCOPES constant with a property that wraps dynamic and
static requirements.

The login command is extended with a `--gcs {id}` option which
requests the relevant manage_collections scope from Auth. To do so, it
adds the GCS to the dynamic requirements of the login manager, then
proceeds with the login flow as normal.

These features are exercised in the addition of a `globus collection
delete` command which operates via a GCSClient to delete a collection.