Refine error handling for 'collection delete'
And add tests for this functionality.
`collection delete` now "remaps" to `endpoint delete` if the detected
endpoint type satisfies the requirements for this command.

The implementation for now is a simple dict which maps endpoint types
to the preferred command to run.

Unrelated changes which were included to aid testing:
- skip-covered in coverage reporting
- add pytest-timeout to quick-fail tests which take too long (happens
  with connection errors which loop retries and sleep)