First draft of "Endpointish"
The idea of the Endpointish is to formalize the mapping between GCSv5
Collection commands and types and the inverse mapping from GCSv4/GCP
commands and types. It should handle
    `collection delete` <-> `endpoint delete`
for both of those commmands, and so forth for all similar
cross-referencing commands on one type or the other.

To support this, some work was needed on the exception handling hooks
to support running on non-GlobusAPIError error classes. In order to
facilitate this, the error handling has been split from the rest of
the `parsing` subpackage and made its own piece.

Unfortunately, this currently introduces a cycle, as
`services.transfer` wants `EXPLICIT_NULL` from the `parsing`
subpackage. Proper resolution of those cycles is deferred here (for
now), but speaks to a deeper structural problem.