Remap 'endpoint delete' to 'collection delete'
Add the mapping of endpoint->collection to the Endpointish error
handling, and use Endpointish to check the type of inputs to
`globus endpoint delete`, the same way that it's used in
`collection delete`.

In support of this, some minor, but helpful, changes:
- give `EndpointType` helpers for getting "standard" tuples of values
- allow the caller to pass a TransferClient to Endpointish, to avoid
  an extra instantiation
- add ExpectedEndpointError, mirroring ExpectedCollectionError

The new behavior is tested on all three of the object types which are
not supported by endpoint delete.