Add testing for 'endpoint permission create'
The ACL creation command was untested and features a variety of
interesting codepaths.

The success cases are fairly simple, but also test the various failure
modes for improper options. In one case, remove a check from the
command itself because it is unreachable -- the checked condition is
handled in the security_principal_opts decorator.

All tests focus on checking exit code and part of output.