Type annotate exception hooks
Annotations in the registry define the various excepthook types.
Notably, we cannot define hooks as operating on `Exception` because
they rely on subclass-specific properties. Instead, they use a type
var with a bound of Exception.

Some bits of typing/mypy weirdness needed workarounds.

You can't declare the type of a variable assigned the empty list
because the type of the right-hand-side mismatches with the
left-hand-side. Solve with a `cast`.

mypy's union elimination rules don't follow all control flow
correctly. In two cases, hook logic was rearranged trivially to make
the fixes.