Add mypy to tox, fixup types
Add mypy to tox as an environment. This lets us run `mypy` easily in
CI without relying on pre-commit file-passing logic. It also changes
the behavior of mypy to run in this way, so it's a good safety check.
`make lint` runs this in addition to pre-commit linting.

The resulting `mypy` invocation fails, mostly because we have
un-annotated functions within the CLI. These are treated as producing
`Any`, so they mess up type-narrowing logic on Optional values.
To solve this, many functions simply need to be annotated as
`Callable -> Callable`.

Add `click` to the requirements for mypy linting in pre-commit. This
lets mypy get types for click objects and methods, rather than
converting everything to Any.