Add 'collection show' command
This pulls in a few significant pieces from gcs-cli without
significant modifications. In particular, the connector name<->id
mapping information is pulled from there (and reformatted a bit), and
the PrincipalResolver (which we will need for other GCS commands) is
pulled in without changes.

The idea of the PrincipalResolver is that it handles rewriting an ID
to a username, or rewriting a URN to a username. We may want to make
this into one of our field objects in the future, but for now it is
left unchanged.

`--include-private-policies` and the test data for it are copied
pretty much verbatim from gcs-cli.

One odd nuance is that if we remap `collection show` and
`endpoint show` the same way we did `delete`, we won't be able to
handle a GCSv5 Host Endpoint. That doesn't make a ton of sense, so
this is tweaked a bit to allow `globus endpoint show` on any
non-collection endpoint(ish).

Tests exercise most of the new functionality, but a bit of
PrincipalResolver is not being used yet.