Add parsing utilities needed for collection update
This pulls in various tools which were defined as part of gcs-cli, and
includes a variety of tests for them.
Specifically, we will need:
- CommaDelimitedList
- JSONStringOrFile
- StringOrNull
- UrlOrNull
- nullable_multi_callback

These are all now defined. Additionally, we update EXPLICIT_NULL to
have a custom repr and to bool as False.

Tests exercise the various new parsing tools.