Initial version of 'collection update'
Right now there are a few crude elements to this.
In particular, there is no "deep understanding" of which options are
appropriate only to Mapped Collections. That means that we have to
check explicit lists of values in the Guest Collection case. Failure
to do so would result in a TypeError, as unsupported parameters are
passed to a GuestCollectionDocument.

The list of supported options is a strict subset of those supported by
gcs-cli, removing some which rely on policy objects (not yet present
in the SDK).

Because this significantly expands the size of the collection tests,
split them up into several smaller modules.