Improve collection unsupported opt detection
For 'collection update', we can inspect the signature on the document
class we're going to use, and then use this to check the parameters
which will actually be passed.

In the few cases of parameters which are rewritten, they turn out to
be common to both collection types and therefore do not need special
treatment.