Begin the journey towards type-checking coverage
Temporarily enable disallow_untyped_defs, fix many of the errors, then
turn the flag off (comment it out) so that builds will pass.
The commented-out directive could be read as a "statement of intent".

We can start working on getting mypy to check the whole CLI, but
turning on this strict flag would make our builds fail today. Rather
than make *all* of the desirable changes -- many of which are very
simple -- for now, take a sample and get the ball rolling.