Improve performance of GCSv5 operations
Remove a roundtrip to Transfer to get the host endpoint's address.
We've checked, and at present we're guaranteed that the server list
for a collection will be populated with the hostname of the underlying
endpoint. The extra call to Transfer can be very slow in pathological
cases, making the commands hang for several seconds while auth info is
being fetched and checked.

In terms of the concrete change:
- get_gcs_client() now accepts `gcs_address` as an input
- Endpointish now has `get_gcs_address()` which asserts that the
  object is of a valid/relevant type
- internal bits and bobs to support these (e.g. an addition to
  EndpointType)