Make `collection list` error on non-endpoint data
If given a non-GCSv5 Endpoint object, `collection list` should abort
with a WrongEndpointTypeError. This was missed during review but is
definitely the right/desirable behavior.

Some slight tweaks to globus_cli.endpointish make it behave nicer in
this case. (Accept a single type for assert_ep_type, a la `isinstance`,
and provide a better message in this case.)

cc @rudyardrichter