Add `globus search query` command with GET & POST
The command supports two usage modes for `GET` and `POST` query
submission. Unlike the globus-search-cli version of this command,
`--offset` is not supported, in anticipation of eventual use of
built-in pagination.

The other major difference from the globus-search-cli commands is the
fact that `query` is supplied as one command, with `-q` or
`--query-document` as options which switch between GET and POST query
forms.