Initial scriv config + support tooling
Because we have an asciidoc changelog and scriv doesn't *quite* output
things in the right format, add a fixer script which modifies the
content a bit. Since `scriv collect --edit` within tox is dicey, just
run $EDITOR from the makefile (which also means that running without
the editor is just a tox target).

Update release doc and move it into a dedicated RELEASING doc.