Fix `endpoint activate` calls for SDKv3
This updates to SDK v3.0.3 to pick up a bugfix to the type
annotations. With this in place, and a few additions to annotations in
the CustomTransferClient, type checking is more complete. This lets
mypy pick up on the incorrect arguments to `endpoint_activate()`,
which requires that `requirements_data=...` is a named parameter.

Linting under mypy now includes SDK v3.