Fix USER_BASE detection for old virtualenv
Older versions of `virtualenv` put a `site.py` module into the
pythonpath for the virtualenv they generate. As a result, the stdlib
`site` module is replaced with its spooky cousin, which has *almost*
the right attributes, but not quite.

To handle, we can't use `site.getuserbase()`. Instead, we'll check if
`site` has `getuserbase()`. If it doesn't, we will immediately fail
the "is user install?" question.
In this case, it's extremely unlikely that we would consider the
install a match for `pip install --user`. The only even possible case
would be a user with a custom `site.py` in their user packages.

As a precaution against mistakes in this logic or the pipx-home logic,
these are now deferred from import-time until they are needed.