Refactor recursive ls and fix bug
`ls --recursive` failed with a hard error on any empty listings. The
ls on `/mnt` is a good test-case for this, but it's also possible to
get this situation easily by applying a filter to a recursive `ls`.

In order to fix the bug, an initial failing test was added. However,
this work also refactors recursive ls to improve the maintainability
of this feature. The main iteration is decomposed into smaller
units and type annotations are applied.