Remove help text for "--batch" when stdin is a tty
`--batch` printed a help message to stderr when stdin is a tty.
However, the message printed is no longer accurate now that `--batch`
takes a filename.

It's not clear who would use `--batch -` and be surprised by the
wait-on-stdin behavior at this point. Rather than working
speculatively to retain it, which would require inspecting the file
handle returned by click.FileType, just drop the help text.

resolves #581