Fix 'globus session consent' crashfail after login
At the end of the login flow, `globus session consent` would expect
`id_token` in the response data in order to do the
no-account-switching validation step we added in v3.0. However, the
`id_token` is only present on login flows which include the `openid`
scope.

To rectify:
- add `LoginManager.always_required_scopes`, an iterator which yields
  the openid scope string, "openid"
- `LoginManager.run_login_flow` adds `always_required_scopes` to any
  scope list (implicit or provided)
- `globus session consent` now passes scopes as a keyword argument,
  not under `session_params`
- annotate and type-check `globus session consent`

This was tested manually, and it is not clear how effective a mocked
test with synthetic response data would be in preventing errors of
this kind in the future.