Fix login always_required_scopes
Expand the set of scopes for `always_required_scopes` (used in
`session consent`) to all of the Auth scopes we request. The reason
for this is that we need openid in order to get `id_token`, but we
need all of the other scopes to avoid "downgrading" the capabilities
of the Auth token we have in storage.

Without this fix, `globus session consent openid; globus whoami`
fails, as does `globus logout`, as the response payload from
`oauth2_userinfo()` changes in the absence of the email and profile
scopes.