Add 'cli-profile-list' command + minor login fix
Testing this work uncovered a minor issue with login, in which client
creds do not trigger a specific message and error. This could be
confusing. Rectify with a custom usage error.

To implement `globus cli-profile-list`, directly check the token
storage database to list the namespaces which are in use. Filter to
those created by the CLI, and then parse them (slash-delimited) to
identify them by type (user vs client) and environment.
Listing is done by table-printing the results.

Tests need to carefully inject data into the test token storage's
in-memory DB to simulate the presence of additional profiles. Once
done, several cases are tested in simple fashion, ensuring that
expected profile names are or are not present in the output.

The `--all` option, which expands the listing to all environments, is
kept hidden for now, on the assumption that it would confuse most
users. If it is frequently used or the fact that it is hidden leads to
confusion, it can be made public later.