cli-profile-list shows default, current profiles
Include the default profile in the output always, even when `--all` is
not given. `--all` is now primarily an environment filter.

Include a small indicator arrow on the row where the current profile
is found. To improve output, the table printer is updated to allow for
a headerless column with no header separator. This is too complex with
the format string which was previously used, so rewrite as a simple
function which produces strings.