Add Globus Search index role commands
Add commands for role create, list, delete, and tests for these
commands with appropriate test fixture data as needed.

Like Transfer ACL commands, these commands support some interactions
with Globus Auth. Response rendering is done with Auth to get
usernames as appropriate for text output, and usernames are accepted
as inputs.

The URN usages here may fit in a new param type in the future, but
none is added yet, as this is the first potential use-case. Instead,
existing helpers are used as much as is possible.