Use pypa's `build` to build dists
`build` replaces use of `setup.py sdist bdist_wheel`. It has several
advantages over the older approach of using `setuptools` and `wheel`
directly:
- it builds in an isolated environment (by default), so that it can't
  be confused by local files when building a wheel
- it can support alternative build backends via pyproject.toml (if we
  add these), per PEP517 [1]
- it's the more modern (and therefore documented) way of doing a build

A pyproject.toml is not needed for this change. `build` touts itself
as an implementation of the standard set out in PEP517, and the spec
states:

  If the pyproject.toml file is absent, or the build-backend key is
  missing, the source tree is not using this specification, and tools
  should revert to the legacy behaviour of running setup.py

This is our case, so `build` works as expected and desired.

[1] https://www.python.org/dev/peps/pep-0517/