Update 'responses' and begin removal of fixed IDs
Tests are currently using fixed strings for a known task_id, and the
IDs of go#ep1 and go#ep2 . It's fine to have test constants, but we
have a new method in which the constants can be exposed to the tests
as loaded response metadata.

Update the test fixture handling to avoid using these known variables
for path templating, but do not eliminate the endpoint fixtures yet
(this will probably touch too many tests).

A few tests are updated to use the sdk _testing registry.