Add 'globus group [join|leave]' commands
Support join, leave, and request_join operations.

For `request_join`, the command is `group join --request`.

For 'leave', the user's active memberships are fetched by default. For
join/request_join, the user's primary ID is used by default. In both
cases, `--identity` is available as a flag to force the operation to
work on a single specific identity.