Cleanup linting config / dependencies
This started with the addition of `types-jmespath` (stubs from
typeshed) and removal of `mypy.ignore_missing_imports`.

However, we have linting tools listed in the `dev` extra. Keeping
these in sync with `.pre-commit-config.yaml` is a manual job and the
purpose is no longer clear since the advent of pre-commit. The
pre-commit config has correct (and enforced) tool versions.
For now, just remove the lint tools from `dev`. Solutions (e.g.
`pip-compile-multi`) can be pursued if easy developer installation of
these tools is or becomes a priority.